Restructure to 'status certificate' instead of 'certificate status'
