+++ cmd/ctl/BUILD.bazel
@@ -32,10 +32,10 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//cmd/ctl/cmd:all-srcs",
-        "//cmd/ctl/pkg/certificate:all-srcs",
         "//cmd/ctl/pkg/convert:all-srcs",
         "//cmd/ctl/pkg/create:all-srcs",
         "//cmd/ctl/pkg/renew:all-srcs",
+        "//cmd/ctl/pkg/status:all-srcs",
         "//cmd/ctl/pkg/version:all-srcs",
     ],
     tags = ["automanaged"],
