+++ cmd/ctl/pkg/status/certificate/BUILD.bazel
@@ -2,8 +2,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["status.go"],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/certificate/status",
+    srcs = ["certificate.go"],
+    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/status/certificate",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/certmanager/v1alpha2:go_default_library",
